+++ Sources/BraintreeCore/BTClientMetadata.swift
@@ -12,7 +12,7 @@ import Foundation
 /// used. In this case, the source and integration may change over time, while
 /// the sessionID should remain constant.
 @_documentation(visibility: private)
-@objcMembers public class BTClientMetadata: NSObject {
+public class BTClientMetadata: Encodable {
 
     /// :nodoc: This property is exposed for internal Braintree use only. Do not use. It is not covered by Semantic Versioning and may change or be removed at any time.
     /// Integration type
@@ -24,8 +24,13 @@ import Foundation
 
     /// :nodoc: This property is exposed for internal Braintree use only. Do not use. It is not covered by Semantic Versioning and may change or be removed at any time.
     /// Auto-generated UUID
-    public var sessionID: String
+    public var sessionID = UUID().uuidString.replacingOccurrences(of: "-", with: "")
     
+    public var platform = "iOS"
+    
+    public var version = BTCoreConstants.braintreeSDKVersion
+    
+    // TODO: - Remove once all POSTs moved to Codable
     /// Additional metadata parameters
     var parameters: [String: Any] {
         [
@@ -37,10 +42,25 @@ import Foundation
         ]
     }
     
-    override init() {
-        self.integration = .custom
-        self.source = .unknown
-        self.sessionID = UUID().uuidString.replacingOccurrences(of: "-", with: "")
-        super.init()
+    private enum CodingKeys: String, CodingKey {
+        case integration = "integration"
+        case sessionID = "sessionId"
+        case source = "source"
+        case platform = "platform"
+        case version = "version"
+    }
+    
+    init(integration: BTClientMetadataIntegration = .custom, source: BTClientMetadataSource = .unknown) {
+        self.integration = integration
+        self.source = source
+    }
+    
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try container.encode(integration.stringValue, forKey: .integration)
+        try container.encode(sessionID, forKey: .sessionID)
+        try container.encode(source.stringValue, forKey: .source)
+        try container.encode(platform, forKey: .platform)
+        try container.encode(version, forKey: .version)
     }
 }