+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## unreleased
 * Add BraintreeShopperInsights module (BETA)
 
+## 6.11.0 (2023-12-20)
+* Update all SDK errors to be public and [Equatable](https://developer.apple.com/documentation/swift/equatable) (fixes #1152 and #1080)
+* BraintreeThreeDSecure
+  * Fix bug where `BTThreeDSecureClient.initializeChallenge()` callback wasn't properly invoked (fixes #1154)
+
 ## 6.10.0 (2023-11-17)
 * BraintreePayPalNativeCheckout
   * Update PayPalCheckout from 1.1.0 to 1.2.0.